requirements.txt
python-telegram-bot==21.6
sqlalchemy
requests
beautifulsoup4
python-dotenv
pdfplumber
psutil
asyncpg
greenlet
httpx==0.27.0